add used items on top

+++ src/main/java/com/example/project2/entities/PauseButton.java
@@ -4,7 +4,7 @@ import javafx.scene.image.Image;
 import javafx.scene.input.KeyEvent;
 
 public class PauseButton extends Entity {
-    public PauseButton(float x, float y, Image img) { super(x, y, img); }
+    public PauseButton(double x, double y, Image img) { super(x, y, img); }
 
     @Override
     public void update() {
